// src/app/features/pokemon/types.ts
export interface PokemonListResponse {
    count: number;
    next: string | null;
    previous: string | null;
    results: { name: string; url: string }[];
  }
  
  export interface PokemonDetail {
    id: number;
    name: string;
    sprites: {
      front_default: string;
      other?: {
        'official-artwork'?: { front_default: string };
      };
    };
    types: { slot: number; type: { name: string } }[];
    abilities: { ability: { name: string } }[];
    height: number;
    weight: number;
  }
  